fix: Place contrastColors in core-utils
