+++ packages/core-utils/src/index.ts
@@ -1,3 +1,4 @@
+import * as contrastColors from "./contrast-colors";
 import * as itinerary from "./itinerary";
 import * as map from "./map";
 // @ts-expect-error not typed
@@ -14,6 +15,7 @@ import * as queryGen from "./query-gen";
 import SafeSuspense from "./suspense";
 
 const core = {
+  contrastColors,
   itinerary,
   map,
   profile,
